Update localStorage.get, remove variables.js, refactored variables, refactor displayTasks and createTasks (input)
